test(frontend): consolidate SendModal specs into SendModal.spec.ts
Move the steps-reactivity tests into a proper SendModal.spec.ts file
under a top-level describe('SendModal') with a nested
describe('steps derivation reactivity (regression coverage)') block,
matching the convention used by other component spec files in this
repo (one spec file per component, nested describe blocks for
sub-areas).

- Create src/frontend/src/tests/lib/components/send/SendModal.spec.ts
  with the SendModal > steps derivation reactivity describe tree.
- Remove the standalone SendModal.steps-reactivity.spec.ts.
- Update the inline comment pointer in SendModal.svelte to reference
  the consolidated spec file and the named describe block.
- No behavioural change; the three reactivity assertions still pass
  (3/3) and ESLint/Prettier are clean.
